move rename cmd tests to own module
